Only add referenced responses to the spec
The process of adding responses dynamically to endpoints is relying on
adding the HTTP Status name as the response instead of an dict
referencing a schema.  This causes marsmallow to interpret it as a
reference so the reference has to exist.

Instead of adding all the possible responses to the spec, only the
responses referenced in any endpoints are added when necessary.  This
reduces noise in the spec while keeping the feature intact.

Fixes #206